integration: move query helpers from tablets to utils
The following helpers are all usable universally, not only for tablets:
- send_statement_everywhere,
- send_unprepared_query_everywhere,

so it makes sense to extract them to utils.rs.
